+++ samples/react-native/metro.config.js
@@ -15,7 +15,7 @@ const monorepoPackages = {
 
 /**
  * Metro configuration
- * https://facebook.github.io/metro/docs/configuration
+ * https://reactnative.dev/docs/metro
  *
  * @type {import('metro-config').MetroConfig}
  */